Add files via upload
Unidentified flying objects (UFOs) have been spotted around the world. Who are they? Where did they come from? These are still open questions. Nevertheless, this project tried to tackle this issue by classifying these objects. One of the convenient tools is machine learning (ML) methods because they are capable of accurately performing such kinds of tasks. A suitable candidate of ML methods is the k-nearest neighbours algorithm as it is simple and requires a relatively small amount of time or epochs to train an ML model. Following this, our project used such an ML algorithm to forecast where and which shape UFOs appear. Another essential feature of our approach is advanced data visualisation via Google Maps. By incorporating the library ``GMapPlot'', we were able to draw a map of UFOs' appearances around the globe, including North America, Europe, and Australia. As the first step of the data analysis, we plotted the correlation matrix and correlation diagram. After that, we performed a proper cross-validation procedure. The next step was the analysis of features where the Natural Language Toolkit (NLTK) was applied. In particular, we came up with a function which calculates the frequency of the words in the description of the observation that was a part of our input parameters to the ML model. Inasmuch as we normalised the inputs during data processing, our further research exploited a logistic regression, whose usage aimed to predict the shape of the UFO based on several parameters, such as a geographical location, day, time, and frequencies of the words in the description. As a result, the obtained model guessed several most frequent shapes with high accuracy and facilitated UFO's detection.

In conclusion, a highly accurate UFOs' classification is of great practical importance for ufologists. Using the implemented ML model, they could determine any type of UFO by filming it and measuring the required parameters.